requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-pptx==1.0.2
python-multipart==0.0.12
pydantic==2.10.3
google-generativeai
python-dotenv
